refactor: Use signed integers

+++ cxx/concurrent.cpp
@@ -8,11 +8,11 @@ int main() {
     std::cerr << "std::thread::hardware_concurrency(): "
               << concurrency << std::endl;
 
-    std::vector<std::future<size_t>> futures;
-    wtl::Semaphore semaphore(2u);
-    constexpr size_t n = 6u;
+    std::vector<std::future<int>> futures;
+    wtl::Semaphore semaphore(2);
+    constexpr int n = 6;
 
-    for (size_t i=0; i<n; ++i) {
+    for (int i = 0; i < n; ++i) {
         // manual lock before thread creation: thread IDs are reused
         semaphore.lock();
         futures.push_back(std::async(std::launch::async, [&semaphore, i] {
@@ -31,7 +31,7 @@ int main() {
     std::cerr << std::endl;
 
     futures.clear();
-    for (size_t i=0; i<n; ++i) {
+    for (int i = 0; i < n; ++i) {
         // RAII lock after thread creation: new thread IDs are created
         futures.push_back(std::async(std::launch::async, [&semaphore, i] {
             std::lock_guard<wtl::Semaphore> _(semaphore);
